+++ mesos/heartbeat.go
@@ -64,3 +64,12 @@ func (e *Scheduler) Heartbeat() {
 		e.Config.Suppress = true
 	}
 }
+
+// HeartbeatLoop - The main loop for the hearbeat
+func (e *Scheduler) HeartbeatLoop() {
+	ticker := time.NewTicker(time.Second * 1)
+	defer ticker.Stop()
+	for ; true; <-ticker.C {
+		e.Heartbeat()
+	}
+}